+++ src/main/java/WGU/Capstone/ChirpCharts/dao/ChirpChatRepository.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 import java.util.List;
 
 @CrossOrigin("http://localhost:4200")
-public interface ChirpChatRepository extends JpaRepository<ChirpChat, Long> {
+public interface ChirpChatRepository extends JpaRepository<ChirpChat, Integer> {
 
     @Query("SELECT c FROM ChirpChat c WHERE c.franchiseId = :franchiseId AND c.year = :year")
     List<ChirpChat> getChatsByTeamAndYear(@Param("franchiseId") Integer franchiseId, @Param("year") Integer year);
